fix(getList): rework of the internals
The logic should now be much cleaner to read and faster to process by the compiler. This allows the usage of aliases on "simple value keys" and not only on pathes. However one resulting change/bugfix was implemented that COULD break something: getList() now returns an array of $default if a single value key was required that does not exist in the input list. See ArrayPathsTest.php:591-591