feat: replace traits with abstract classes and inheritance
I'm sorry for the hassle. Traits with own properties are a nightmare to extend, which I found out when migrating my own libraries. Therefore I dropped the traits completely and replaced them with abstract classes that extend each other. As this change is done so short after the release of v3 I don't consider it a "breaking" change, because other than me no one has updated the code to v3 yet. If you did, please don't hate me :(.